split into grouper.py
